fix(startup): the launch card imports the repository walk from git-head
The fs-only repository walk and worktree resolution move into git-head.ts
beside the rest of the from-files answer, so a launch-card module that
imports active-repo-context stops evaluating git.ts's process layer
(node:child_process) it never calls. git.ts composes the moved resolvers
with its subprocess half and keeps its exported surface.
