+++ packages/coding-agent/src/utils/active-repo-context.ts
@@ -21,7 +21,7 @@ import { isEnoent } from "@veyyon/utils/fs-error";
 import * as logger from "@veyyon/utils/logger";
 import { errorMessage } from "@veyyon/utils/type-guards";
 
-import { type GitRepository, repo } from "./git";
+import { type GitRepository, linkedWorktreeSync, resolveRepository, resolveRepositorySync } from "./git-head";
 
 export interface ActiveRepoContext {
 	cwd: string;
@@ -44,7 +44,7 @@ export interface WorktreeContext {
  * resolve to the shared `foo.git` dir, so a trailing `.git` is stripped.
  */
 export function resolveWorktreeContext(cwd: string): WorktreeContext | null {
-	const worktree = repo.linkedWorktreeSync(cwd);
+	const worktree = linkedWorktreeSync(cwd);
 	if (!worktree) return null;
 	const base = path.basename(worktree.primaryRoot);
 	const projectName = base.endsWith(".git") ? base.slice(0, -4) : base;
@@ -98,9 +98,9 @@ function reportUnlistableCwd(cwd: string, error: unknown): void {
 	});
 }
 
-async function resolveRepository(cwd: string): Promise<GitRepository | null> {
+async function resolveRepositoryQuiet(cwd: string): Promise<GitRepository | null> {
 	try {
-		return await repo.resolve(cwd);
+		return await resolveRepository(cwd);
 	} catch {
 		// Null means "cwd is not inside a repository", and this resolution is a filesystem walk up the
 		// directory chain, so a failure at any level means the same thing to the caller: keep looking for a
@@ -109,9 +109,9 @@ async function resolveRepository(cwd: string): Promise<GitRepository | null> {
 	}
 }
 
-function resolveRepositorySync(cwd: string): GitRepository | null {
+function resolveRepositorySyncQuiet(cwd: string): GitRepository | null {
 	try {
-		return repo.resolveSync(cwd);
+		return resolveRepositorySync(cwd);
 	} catch {
 		// Same verdict as the asynchronous twin above.
 		return null;
@@ -217,12 +217,12 @@ function findSingleDirectChildRepoSync(cwd: string): ActiveRepoContext | null {
 
 export async function resolveActiveRepoContext(cwd: string): Promise<ActiveRepoContext | null> {
 	const resolvedCwd = path.resolve(cwd);
-	if (await resolveRepository(resolvedCwd)) return null;
+	if (await resolveRepositoryQuiet(resolvedCwd)) return null;
 	return findSingleDirectChildRepo(resolvedCwd);
 }
 
 export function resolveActiveRepoContextSync(cwd: string): ActiveRepoContext | null {
 	const resolvedCwd = path.resolve(cwd);
-	if (resolveRepositorySync(resolvedCwd)) return null;
+	if (resolveRepositorySyncQuiet(resolvedCwd)) return null;
 	return findSingleDirectChildRepoSync(resolvedCwd);
 }
