+++ packages/coding-agent/src/utils/git-head.ts
@@ -167,6 +167,126 @@ export function resolveRepositorySync(startDir: string): GitRepository | null {
 	}
 }
 
+/**
+ * The same repository walk and worktree resolution, asynchronous. `git.ts`
+ * imports these rather than keeping a second copy, and a caller that must not
+ * reach `git.ts`'s process layer imports them from here.
+ */
+
+export async function retryOnEintr<T>(op: () => Promise<T>): Promise<T | null> {
+	for (let attempt = 0; attempt <= EINTR_MAX_RETRIES; attempt += 1) {
+		try {
+			return await op();
+		} catch (err) {
+			if (shouldRetry(err, attempt)) continue;
+			return null;
+		}
+	}
+	throw new Error("retryOnEintr: exhausted without resolution");
+}
+
+export async function getEntryType(gitEntryPath: string): Promise<EntryType | null> {
+	return retryOnEintr(async () => {
+		const stat = await fs.promises.stat(gitEntryPath);
+		if (stat.isDirectory()) return "directory";
+		if (stat.isFile()) return "file";
+		return null;
+	});
+}
+
+export async function readOptionalText(filePath: string): Promise<string | null> {
+	return retryOnEintr(async () => await fs.promises.readFile(filePath, "utf8"));
+}
+
+async function resolveGitDir(gitEntryPath: string, entryType: EntryType): Promise<string | null> {
+	if (entryType === "directory") return gitEntryPath;
+	const content = await readOptionalText(gitEntryPath);
+	if (content === null) return null;
+	const parsed = parseGitDirPointer(content);
+	if (!parsed) return null;
+	const gitDir = path.resolve(path.dirname(gitEntryPath), parsed);
+	return (await getEntryType(gitDir)) === "directory" ? gitDir : null;
+}
+
+export async function resolveCommonDir(gitDir: string): Promise<string> {
+	const content = await readOptionalText(path.join(gitDir, "commondir"));
+	const relative = content?.trim();
+	if (!relative) return gitDir;
+	return path.resolve(gitDir, relative);
+}
+
+async function resolveRepoFromEntry(
+	repoRoot: string,
+	gitEntryPath: string,
+	entryType: EntryType,
+): Promise<GitRepository | null> {
+	const gitDir = await resolveGitDir(gitEntryPath, entryType);
+	if (!gitDir) return null;
+	return {
+		commonDir: await resolveCommonDir(gitDir),
+		gitDir,
+		gitEntryPath,
+		headPath: path.join(gitDir, "HEAD"),
+		repoRoot,
+	};
+}
+
+export async function resolveRepository(startDir: string): Promise<GitRepository | null> {
+	let current = path.resolve(startDir);
+	while (true) {
+		const gitEntryPath = path.join(current, ".git");
+		const entryType = await getEntryType(gitEntryPath);
+		if (entryType) {
+			const repository = await resolveRepoFromEntry(current, gitEntryPath, entryType);
+			if (repository) return repository;
+		}
+		const parent = path.dirname(current);
+		if (parent === current) return null;
+		current = parent;
+	}
+}
+
+function isLinkedWorktree(repository: GitRepository): boolean {
+	return (
+		repository.gitDir !== repository.commonDir &&
+		getEntryTypeSync(path.join(repository.gitDir, "commondir")) === "file"
+	);
+}
+
+async function isLinkedWorktreeAsync(repository: GitRepository): Promise<boolean> {
+	return (
+		repository.gitDir !== repository.commonDir &&
+		(await getEntryType(path.join(repository.gitDir, "commondir"))) === "file"
+	);
+}
+
+/** Resolve the primary checkout root from an already-resolved repository, synchronously. */
+export function primaryRootFromRepositorySync(repository: GitRepository): string {
+	if (path.basename(repository.commonDir) === ".git") return path.dirname(repository.commonDir);
+	if (isLinkedWorktree(repository)) return repository.commonDir;
+	return repository.repoRoot;
+}
+
+/** The asynchronous twin of {@link primaryRootFromRepositorySync}. */
+export async function primaryRootFromRepository(repository: GitRepository): Promise<string> {
+	if (path.basename(repository.commonDir) === ".git") return path.dirname(repository.commonDir);
+	if (await isLinkedWorktreeAsync(repository)) return repository.commonDir;
+	return repository.repoRoot;
+}
+
+/**
+ * Linked-worktree metadata for `cwd`, or `null` when `cwd` is the primary
+ * checkout (or outside a repository). `root` is the worktree's own checkout
+ * root; `primaryRoot` is the shared main checkout that names the project.
+ * Resolves purely via on-disk `.git`/`commondir` walking — no subprocess —
+ * so the status line may call it on every render.
+ */
+export function linkedWorktreeSync(cwd: string): { root: string; primaryRoot: string } | null {
+	const repository = resolveRepositorySync(cwd);
+	if (!repository || !isLinkedWorktree(repository)) return null;
+	return { root: repository.repoRoot, primaryRoot: primaryRootFromRepositorySync(repository) };
+}
+
 export function getRefLookupDirs(repository: GitRepository): string[] {
 	if (repository.gitDir === repository.commonDir) return [repository.gitDir];
 	return [repository.gitDir, repository.commonDir];
